LIBASPACE-139. Changed search placeholder to "Search Within Collection"
Overrode "actions.search_in" I18N resource from stock ArchivesSpace
using "public/locales/en.yml" file.

Also changed the spacing in the en.yml file to use 2 spaces for
indentation, instead of 4, both to match the style in stock
ArchivesSpace, and Ruby style.

Added CSS to adjust the width of the form group and "Search" textfield
to the full width of the sidebar, so that the placeholder text would
not be truncated.

Note: The change to the "actions.search_in" resource also changes
the HTML title (displayed in a browser tab).

https://issues.umd.edu/browse/LIBASPACE-139