LIBASPACE-138. Moved "Found in" over "Overview" summary in search results
Switched the code so that the "Found in" section in the search results
(which displays the repository name) is displayed before the "Overview"
section (which displays the abstract for the respository).

Note: Needed to add 'style="clear:both"' to the "result_context" class,
in order for the text to be laid out properly.

https://issues.umd.edu/browse/LIBASPACE-138