LIBASPACE-119. Reversed "abstract" and "scopecontext" text in display
Added the "public/views/shared/_record_innards.html.erb" to
"umd-lib-aspace-theme" plugin, and customized so that the "abstract"
field (when available) would be used in preference to the "scopecontext"
field for the "above the fold" description of a resource.

When the "abstract" is used, the "scopecontext" field will be displayed
in the "Additional Description" accordion tab.

https://issues.umd.edu/browse/LIBASPACE-119