LIBASPACE-200. Added "Home" link to navigation bar.
Added "Home" link to navigation bar.

Also, modified the way the "Help" link was added, to better conform
with ArchivesSpace practices. Since the "Help" link is a fully-qualified
URL, had to add some logic into the "_navigation.html.erb" helper to
add the app_prefix only if the URL was relative.

Added "brand.home" to the plugin "en.yml" file to protect against any
changes to the ArchivesSpace version.

https://issues.umd.edu/browse/LIBASPACE-200