LIBASPACE-200. Adjusted Website Feedback link
Adjusted styling on "website feedback" text and link in the header to
match the mockup.

Adjusted the Bootstrap column setting so text/link in header would
align with the right edge of the navigation bar.

Parameterized the actual website feedback URL using AppConfig, as it
appears in two places (in the header and in the footer).

Updated header and footer to use the parameterized website feedback URL.

https://issues.umd.edu/browse/LIBASPACE-200