LIBASPACE-210. Request instructions now display consistently
Modified request instructions for "collections" objects
so that the instructions are displayed as part of the template, not
via Rails "flash" messages.

Added "public/views/shared/_request_instructions.html.erb" as a view
helper that can be added where needed to display the instructions.

Added the view helper to the "show" templates for "objects" and
"resources", which were copied from the v2.2.2 ArchivesSpace codebase.

Modified CSS to style the request instructions so that it appears
similar to the Rails "informational" flash notices (this will likely
be changed in the future).

Updated the wording for the request instructions as specified in
LIBASPACE-211, and made displayed in bold as specified in LIBASPACE-213.

https://issues.umd.edu/browse/LIBASPACE-210