LIBASPACE-117. Navigation search textbox formatting updates
Modified the search textbox in the navigation bar to match mockup at
https://umd-lib.github.io/mockups/aspace/home-letter.htm,
including RWD behavior.

Moved search textbox and button into separate helper, because it is
added twice in the page for RWD handling. Added a local parameter
"search_box_id" that gets passed into the helper to distinguish
the HTML ids used in the form tags.

https://issues.umd.edu/browse/LIBASPACE-117