LIBASPACE-341. Commented out "+" button in search form
Commented out the "+" button in the search form that enables additional
search criteria to be added. This functionality is not working as of
ArchivesSpace v3.1.1, as this UMD-customized template file has not been
updated to the latest version of the template
(see "public/app/views/shared/_search.html.erb").

Updating to the stock ArchivesSpace template (and modifying to support
the UMD customizations to the template) will enable the search criteria
functionality. However, other UMD customizations of the application
cause form field layout issues. We are not prepared at this time to
undertake fixing these layout issues, so simply removing the
functionality is a stop-gap measure.

https://umd-dit.atlassian.net/browse/LIBASPACE-341